cisfun

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int i;
 long int li;
 long long int lli;
 float f;
-printf("Size of a char: %lu byte(s)\n", (long int) sizeof(c));
-printf("Size of an int: %lu byte(s)\n", (long int) sizeof(i));
-printf("Size of a long int: %lu byte(s)\n", (long int) sizeof(li));
-printf("Size of a long long int: %lu byte(s)\n", (long int) sizeof(lli));
-printf("Size of float: %lu byte(s)\n", (long int) sizeof(f));
+printf("Size of a char: %lu byte(s)\n", sizeof(c));
+printf("Size of an int: %lu byte(s)\n", sizeof(i));
+printf("Size of a long int: %lu byte(s)\n", sizeof(li));
+printf("Size of a long long int: %lu byte(s)\n", sizeof(lli));
+printf("Size of float: %lu byte(s)\n", sizeof(f));
 return (0);
 }
